Simplify tests involving multiple datasources as alternatives
